font: fix proportional font rendering for 64-bit
The content of xg_fonadr is read from a file,
and the array type mismatch caused rendering of
proportional fonts to fail on 64-bit builds.

Declare xg_fonadr as int32_t instead of "long int"
to match the file layout in all builds. Remove the
workaround (which only worked for ASCII characters
anyway).

Note that this causes a type mismatch between LOPIF
and the Arachne core. However, LOPIF exclusively
targets DOS where both types are equivalent.

+++ linglue.c
@@ -497,7 +497,7 @@ int xg_fbyt;		      /* Number of bytes in one char of the font */
 int xg_flag;		      /* Types: bit 0 - spec EGA pal */
 			      /*	bit 1 - aktive patt  */
 unsigned char xg_fonlen[256];  // Widths of chars for proportional fonts
-long  int     xg_fonadr[256];  // Beginning of chars for proportional fonts
+int32_t       xg_fonadr[256];  // Beginning of chars for proportional fonts
 unsigned char xg_foncon;       // Flag - const/prop font [0/average]
 unsigned char xg_fonmem;       // Where is the font: 0-MEM(xg_fbuf),1-XMS,2-DISK
 int	      xg_fonhan;       // Handle for XMS/DISK
